Fix incorect sizeof() (jack-session)
The incorrect sizeof resulted in invalid session dirs,
if they were longer than 64 chars, to be passed to clients.

port renames were probably affected as well

The bug was introduced in the jack-session git branch
in d7772d7e5934f019d2ba9f285dea4672053b7d27. It appeared
in svn trunk when the branch was merged in r4057. The first
released version with this bug is 1.9.7

git-svn-id: http://subversion.jackaudio.org/jack/jack2/trunk/jackmp@4538 0c269be4-1314-0410-8aa9-9f06e86f4224