jackdbus: group batch of saves into one
When a parameter is set, jackdbus used to save immediately.
This was quite annoying and ineffective when configuring
lot of parameters without delay. It caused log line to appear
for each parameter set and xml file serialization as well.

This changeset implements delayed save. Parameters are saved
two seconds after last parameter set.

git-svn-id: http://subversion.jackaudio.org/jack/jack2/trunk/jackmp@4583 0c269be4-1314-0410-8aa9-9f06e86f4224