Change JACK_DEFAULT_SELF_CONNECT_MODE to "allow"
The constant was introduced when the mode parameter of the JackEngine
constructor was added in aab13501f44e30636a505dbacfb58dbf08b85340.
Then in 6253344da6021edfdd36bf702f2d97b2e21cb7eb the parameter
was exposed through the control api, but with different default
value.

This commit changes the value to match the default of the control api
parameter.