Allow setting group/permissions of filesystem objects in promiscuous mode
If the value of the JACK_PROMISCOUS_SERVER environment variable is set to the
name or gid of a valid UNIX group for the current user, then all filesystem
objects (e.g. sockets, semaphores and shared memory segments) will be created
with this group as owner, and the permissions will be set to 0660. If not, the
permissions will be set to 0666.
This remove the requirement to use a way too permissive umask for promiscuous
mode to work properly.
For that purpose, a new jack_group2gid helper function has been introduced in
its own obj file so it can be called from both C and C++ code.