Begin JackPosixSemaphore names with slash
According to the posix standard, a named semaphore needs to begin with a
slash in order to guarantee the same semaphore will be opened by that name
by multiple processes.

This is an issue on QNX where a named semaphore that does not begin with a
slash has the current working directory prepended to the name as a kind of
namespacing.