Make JackPortRenameCallback return void, to match JACK1 API JACK2 added this function first, but the int return has been always wrong. When JACK1 added JackPortRenameCallback it used the proper return.
Now that JACK1 supports port renames, devs will start to use it.
(previously it didn't work properly because of the missing jack_client_t* arg)

Some code might be broken because of this, but it's a very simple change,
and existing code would have been broken when changing JACK1 to JACK2.

Finally, JACK2 code never uses the int return value of this callback.
So there's no real reason to NOT change this.