More QNX build fixes
This patch adds a missing case to the tests wscript, to handle the QNX
platform.  It also adds code to the setup for the various directory
macros, to handle the case when the build is taking place on Windows,
to ensure that the directory paths still come across in POSIX format.

Change-Id: If8054284d390e40c97c8c585af2851e94275aad8