jack_connect: Connect all ports with the same alias
If there are multiple ports with the same alias the user has to call
jack_connect() only once and all ports with the same alias will be
connected. With this feature for example a media player does not need to
know which JACK clients are interested in its audio data. The media
player has only to connect to the right alias and will automatically be
connected to all ports with the same alias.

Without this patch only the first register port with the matching name
or alias will be connected. The following ports were ignored.

Signed-off-by: Timo Wischer <twischer@de.adit-jv.com>