JackPosixProcessSync::LockedTimedWait: Check for time jumps
POSIX functions with names ending with _timedwait use the real time
clock and as such as sensitive to changes in system time during the
call. This is described in: [1].

This patch implements a wrapper for pthread_cond_timedwait() that, if
the function call fails, will check if the failure is due to a time warp
and if so, will restart the call.

Based on idea and implementation by Ben Hutchings <ben.hutchings@codethink.co.uk>

[1] https://sourceware.org/ml/libc-alpha/2018-12/msg00512.html

v2: Check if CLOCK_MONOTONIC is available.

Signed-off-by: Maciej Wolny <maciej.wolny@codethink.co.uk>