Move jack_control to dbus (#831)
* Move jack_control to dbus dir

dbus/jack_control:
The `jack_control` script is dbus specific and not part of the original
jack tools. As the files offered by jack-example-tools are now
optionally not built, the script needs to be made available from a
location, that is not ignored when omitting the build of
jack-example-tools files.

* Move installation of jack_control to dbus integration

dbus/wscript:
Install `jack_control` script if building/installing jackdbus.

tools/wscript:
Remove (unconditional) installation of `jack_control` script.