Remove unused JackArgParser::GetArgv
This is not called anywhere. Also this function doesn't do anything useful besides
leaking memory in its current form. I think the idea was to give this a
`char ***` (instead of `char **`) paramter that will be filled and then can
be used by the caller. But in its current form it just allocates a pointer
with the address being only used in the local variable (which then goes out
of scope and the memory is lost).